Refactoring
The validation process for urls will be moved to when they are inputed,
not when they are read from the watchlist. The URL module implements a
function where such actions can be inserted.

Also, a parameter for the supervisor pid has been added to allow Brokers
to send a message when they are done.